+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## [1.9.1]
+* Cursor jumps to the most appropriate offset to display selection.
+
+## [1.9.0]
+* Support inline image.
+
+## [1.8.3]
+* Updated quill_delta.
+
+## [1.8.2]
+* Support mobile image alignment.
+
 ## [1.8.1]
 * Support mobile custom size image.